restruct generic
